Migrate Realtime Crypto test
