+++ .jshintrc
@@ -8,6 +8,7 @@
     "setTimeout",
     "global",
     "window",
+    "isBrowser",
     "define",
     "require",
     "module",
