+++ spec/common/modules/shared_helper.js
@@ -3,30 +3,33 @@
 /* Shared test helper for the Jasmine test suite that simplifies
    the dependencies by provoding common methods in a single dependency */
 
-define(['spec/common/modules/testapp_module', 'spec/common/modules/client_module'], function(testAppModule, clientModule) {
-  var displayError = function(err) {
-    if(typeof(err) == 'string')
-      return err;
+define(['spec/common/modules/testapp_module', 'spec/common/modules/client_module', 'spec/common/modules/testdata_module'],
+  function(testAppModule, clientModule, testDataModule) {
+    var displayError = function(err) {
+      if(typeof(err) == 'string')
+        return err;
 
-    var result = '';
-    if(err.statusCode)
-      result += err.statusCode + '; ';
-    if(typeof(err.message) == 'string')
-      result += err.message;
-    if(typeof(err.message) == 'object')
-      result += JSON.stringify(err.message);
+      var result = '';
+      if(err.statusCode)
+        result += err.statusCode + '; ';
+      if(typeof(err.message) == 'string')
+        result += err.message;
+      if(typeof(err.message) == 'object')
+        result += JSON.stringify(err.message);
 
-    return result;
-  };
+      return result;
+    };
 
-  return module.exports = {
-    setupApp:     testAppModule.setup,
-    tearDownApp:  testAppModule.tearDown,
-    getTestApp:   testAppModule.getTestApp,
+    return module.exports = {
+      setupApp:     testAppModule.setup,
+      tearDownApp:  testAppModule.tearDown,
+      getTestApp:   testAppModule.getTestApp,
 
-    AblyRest:     clientModule.AblyRest,
-    AblyRealtime: clientModule.AblyRealtime,
+      AblyRest:     clientModule.AblyRest,
+      AblyRealtime: clientModule.AblyRealtime,
 
-    displayError: displayError
-  };
-});
+      loadTestData: testDataModule.loadTestData,
+
+      displayError: displayError
+    };
+  });
