+++ spec/realtime/assets/crypto-data-128.json
