+++ spec/realtime/assets/crypto-data-256.json
